+++ demo/ExpenseTracker/AspNetCore/Program.cs
@@ -21,6 +21,6 @@ app.UseVmsShellStaticFiles();
 
 app.MapControllers();
 
-app.MapFallbackToFile("index.html");
+app.MapVmsShellFallbackToFile("index.html");
 
 app.Run();